refactor(oci-collector): parallelize collecting artifacts, establish collector interface
